Apply various IDE suggestions
Several places in the code did not proprely utilize
the posibility of a field to be final. Additionally
the non-generic access in Shared Resource is fixed.
